debug empty divs not fading out after Task
This was probably due to having too many transitions and
animations/flips at the end of the Task component, so I moved the
block state resetting loop to the beginning of the Quiz component.
The associated block transitions should hopefully occur before the
Quiz component fades into the user's view.