again debug empty divs not fading out (for Quiz teaching ex)
probably same problem as last
time (https://github.com/chen10an/active-blicket-comp/commit/c01b1416f85a668e35704de6133ba6b342986462):
too many blocks transitioning out when continuing from one component
to another

OR maybe the problem this time is that I put transitions on Blocks that don't
need transitions, i.e., they have a
"start" position but don't need to go to any "end" point

because the problem was resolved by removing transitions on the
blickets/nonblickets that just act as buttons for their piles and don't need any
transitions. But it was ok to keep the transitions for all the
blickets/nonblickets (there are lots of them!) that need to transition on/off the detector.